requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
oslo.concurrency # Apache-2.0
oslo.config!=4.3.0,!=4.4.0 # Apache-2.0
oslo.log # Apache-2.0
testtools # MIT
pbr!=2.1.0 # Apache-2.0
eventlet!=0.18.3,!=0.20.1,!=0.21.0,!=0.23.0 # MIT
keystoneauth1 # Apache-2.0
oslotest # Apache-2.0
PyYAML # MIT

masakari-monitors
parallel-ssh